# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Media Suite API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface UploadResponse {
  url: string;
}

export interface ImageGenerationRequest {
  modelId: string;
  prompt: string;
  negativePrompt?: string;
  aspectRatio?: string;
  /**
   * @minimum 1
   * @maximum 4
   */
  numImages?: number;
  seed?: number;
  /** Source image URL for image-to-image models */
  imageUrl?: string;
}

export interface VideoGenerationRequest {
  modelId: string;
  prompt: string;
  negativePrompt?: string;
  duration?: string;
  aspectRatio?: string;
  imageUrl?: string;
  seed?: number;
  generateAudio?: boolean;
}

export type MotionGenerationRequestCharacterOrientation =
  (typeof MotionGenerationRequestCharacterOrientation)[keyof typeof MotionGenerationRequestCharacterOrientation];

export const MotionGenerationRequestCharacterOrientation = {
  video: "video",
  image: "image",
} as const;

export interface MotionGenerationRequest {
  modelId: string;
  /** Character/subject image URL */
  imageUrl: string;
  /** Reference motion video URL */
  videoUrl: string;
  prompt?: string;
  characterOrientation?: MotionGenerationRequestCharacterOrientation;
}

export interface LipsyncGenerationRequest {
  modelId: string;
  /** Face video URL */
  videoUrl: string;
  /** Audio to sync URL */
  audioUrl: string;
}

export interface TtsGenerationRequest {
  modelId: string;
  text: string;
  voice?: string;
  speed?: number;
  /** Reference voice audio URL for voice cloning */
  referenceAudioUrl?: string;
}

export type GenerationResponseStatus =
  (typeof GenerationResponseStatus)[keyof typeof GenerationResponseStatus];

export const GenerationResponseStatus = {
  queued: "queued",
  in_progress: "in_progress",
  completed: "completed",
  failed: "failed",
} as const;

export type GenerationResponseMediaType =
  (typeof GenerationResponseMediaType)[keyof typeof GenerationResponseMediaType];

export const GenerationResponseMediaType = {
  image: "image",
  video: "video",
  audio: "audio",
} as const;

export interface MediaOutput {
  url: string;
  contentType?: string;
  width?: number;
  height?: number;
  duration?: number;
}

export interface GenerationResponse {
  jobId: string;
  requestId?: string;
  status: GenerationResponseStatus;
  mediaType: GenerationResponseMediaType;
  outputs?: MediaOutput[];
  error?: string;
}

export type JobStatusResponseStatus =
  (typeof JobStatusResponseStatus)[keyof typeof JobStatusResponseStatus];

export const JobStatusResponseStatus = {
  queued: "queued",
  in_progress: "in_progress",
  completed: "completed",
  failed: "failed",
} as const;

export interface JobStatusResponse {
  jobId: string;
  requestId?: string;
  status: JobStatusResponseStatus;
  queuePosition?: number;
  outputs?: MediaOutput[];
  error?: string;
  logs?: string[];
}

export type ModelInfoCategory =
  (typeof ModelInfoCategory)[keyof typeof ModelInfoCategory];

export const ModelInfoCategory = {
  image: "image",
  video: "video",
  "motion-control": "motion-control",
  lipsync: "lipsync",
  tts: "tts",
} as const;

export interface ModelInfo {
  id: string;
  name: string;
  provider: string;
  category: ModelInfoCategory;
  type: string;
  description?: string;
  tags?: string[];
  supportsDuration?: boolean;
  supportsImageInput?: boolean;
  supportsVideoInput?: boolean;
  supportsAudioInput?: boolean;
  maxDuration?: number;
  aspectRatios?: string[];
  voices?: string[];
}

export interface ModelsListResponse {
  models: ModelInfo[];
}

export type UploadFileBody = {
  file: Blob;
};

export type GetJobStatusParams = {
  requestId: string;
  modelId: string;
};