# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: AI Media Suite API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: generate
  - name: jobs
  - name: upload
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /upload:
    post:
      operationId: uploadFile
      tags: [upload]
      summary: Upload a file to fal.ai storage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "200":
          description: Uploaded file URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /generate/image:
    post:
      operationId: generateImage
      tags: [generate]
      summary: Generate image using AI model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageGenerationRequest"
      responses:
        "200":
          description: Job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /generate/video:
    post:
      operationId: generateVideo
      tags: [generate]
      summary: Generate video using AI model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoGenerationRequest"
      responses:
        "200":
          description: Job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /generate/motion:
    post:
      operationId: generateMotion
      tags: [generate]
      summary: Motion control video generation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MotionGenerationRequest"
      responses:
        "200":
          description: Job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /generate/lipsync:
    post:
      operationId: generateLipsync
      tags: [generate]
      summary: Lip sync video to audio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LipsyncGenerationRequest"
      responses:
        "200":
          description: Job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /generate/tts:
    post:
      operationId: generateTts
      tags: [generate]
      summary: Text to speech
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TtsGenerationRequest"
      responses:
        "200":
          description: Job started or completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /jobs/{jobId}/status:
    get:
      operationId: getJobStatus
      tags: [jobs]
      summary: Get job status and result
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
        - name: requestId
          in: query
          required: true
          schema:
            type: string
        - name: modelId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatusResponse"

  /models:
    get:
      operationId: listModels
      tags: [generate]
      summary: List available AI models
      responses:
        "200":
          description: Models list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelsListResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]

    UploadResponse:
      type: object
      properties:
        url:
          type: string
      required: [url]

    ImageGenerationRequest:
      type: object
      properties:
        modelId:
          type: string
        prompt:
          type: string
        negativePrompt:
          type: string
        aspectRatio:
          type: string
        numImages:
          type: integer
          minimum: 1
          maximum: 4
        seed:
          type: integer
        imageUrl:
          type: string
          description: Source image URL for image-to-image models
      required: [modelId, prompt]

    VideoGenerationRequest:
      type: object
      properties:
        modelId:
          type: string
        prompt:
          type: string
        negativePrompt:
          type: string
        duration:
          type: string
        aspectRatio:
          type: string
        imageUrl:
          type: string
        seed:
          type: integer
        generateAudio:
          type: boolean
      required: [modelId, prompt]

    MotionGenerationRequest:
      type: object
      properties:
        modelId:
          type: string
        imageUrl:
          type: string
          description: Character/subject image URL
        videoUrl:
          type: string
          description: Reference motion video URL
        prompt:
          type: string
        characterOrientation:
          type: string
          enum: ["video", "image"]
      required: [modelId, imageUrl, videoUrl]

    LipsyncGenerationRequest:
      type: object
      properties:
        modelId:
          type: string
        videoUrl:
          type: string
          description: Face video URL
        audioUrl:
          type: string
          description: Audio to sync URL
      required: [modelId, videoUrl, audioUrl]

    TtsGenerationRequest:
      type: object
      properties:
        modelId:
          type: string
        text:
          type: string
        voice:
          type: string
        speed:
          type: number
        referenceAudioUrl:
          type: string
          description: Reference voice audio URL for voice cloning
      required: [modelId, text]

    GenerationResponse:
      type: object
      properties:
        jobId:
          type: string
        requestId:
          type: string
        status:
          type: string
          enum: ["queued", "in_progress", "completed", "failed"]
        mediaType:
          type: string
          enum: ["image", "video", "audio"]
        outputs:
          type: array
          items:
            $ref: "#/components/schemas/MediaOutput"
        error:
          type: string
      required: [jobId, status, mediaType]

    JobStatusResponse:
      type: object
      properties:
        jobId:
          type: string
        requestId:
          type: string
        status:
          type: string
          enum: ["queued", "in_progress", "completed", "failed"]
        queuePosition:
          type: integer
        outputs:
          type: array
          items:
            $ref: "#/components/schemas/MediaOutput"
        error:
          type: string
        logs:
          type: array
          items:
            type: string
      required: [jobId, status]

    MediaOutput:
      type: object
      properties:
        url:
          type: string
        contentType:
          type: string
        width:
          type: integer
        height:
          type: integer
        duration:
          type: number
      required: [url]

    ModelInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        provider:
          type: string
        category:
          type: string
          enum: ["image", "video", "motion-control", "lipsync", "tts"]
        type:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        supportsDuration:
          type: boolean
        supportsImageInput:
          type: boolean
        supportsVideoInput:
          type: boolean
        supportsAudioInput:
          type: boolean
        maxDuration:
          type: integer
        aspectRatios:
          type: array
          items:
            type: string
        voices:
          type: array
          items:
            type: string
      required: [id, name, provider, category, type]

    ModelsListResponse:
      type: object
      properties:
        models:
          type: array
          items:
            $ref: "#/components/schemas/ModelInfo"
      required: [models]