# lib/api-zod/src/generated/types/generationResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Media Suite API
 * OpenAPI spec version: 0.2.0
 */
import type { GenerationResponseMediaType } from "./generationResponseMediaType";
import type { GenerationResponseStatus } from "./generationResponseStatus";
import type { MediaOutput } from "./mediaOutput";

export interface GenerationResponse {
  jobId: string;
  requestId?: string;
  status: GenerationResponseStatus;
  mediaType: GenerationResponseMediaType;
  outputs?: MediaOutput[];
  error?: string;
}