# lib/api-zod/src/generated/types/imageGenerationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Media Suite API
 * OpenAPI spec version: 0.2.0
 */

export interface ImageGenerationRequest {
  modelId: string;
  prompt: string;
  negativePrompt?: string;
  aspectRatio?: string;
  /**
   * @minimum 1
   * @maximum 4
   */
  numImages?: number;
  seed?: number;
  /** Source image URL for image-to-image models */
  imageUrl?: string;
}