# lib/api-zod/src/generated/types/jobStatusResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Media Suite API
 * OpenAPI spec version: 0.2.0
 */
import type { JobStatusResponseStatus } from "./jobStatusResponseStatus";
import type { MediaOutput } from "./mediaOutput";

export interface JobStatusResponse {
  jobId: string;
  requestId?: string;
  status: JobStatusResponseStatus;
  queuePosition?: number;
  outputs?: MediaOutput[];
  error?: string;
  logs?: string[];
}