# lib/api-zod/src/generated/types/lipsyncGenerationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Media Suite API
 * OpenAPI spec version: 0.2.0
 */

export interface LipsyncGenerationRequest {
  modelId: string;
  /** Face video URL */
  videoUrl: string;
  /** Audio to sync URL */
  audioUrl: string;
}