# lib/api-zod/src/generated/types/mediaOutput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Media Suite API
 * OpenAPI spec version: 0.2.0
 */

export interface MediaOutput {
  url: string;
  contentType?: string;
  width?: number;
  height?: number;
  duration?: number;
}