# lib/api-zod/src/generated/types/modelInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Media Suite API
 * OpenAPI spec version: 0.2.0
 */
import type { ModelInfoCategory } from "./modelInfoCategory";

export interface ModelInfo {
  id: string;
  name: string;
  provider: string;
  category: ModelInfoCategory;
  type: string;
  description?: string;
  tags?: string[];
  supportsDuration?: boolean;
  supportsImageInput?: boolean;
  supportsVideoInput?: boolean;
  supportsAudioInput?: boolean;
  maxDuration?: number;
  aspectRatios?: string[];
  voices?: string[];
}