# lib/api-zod/src/generated/types/modelsListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Media Suite API
 * OpenAPI spec version: 0.2.0
 */
import type { ModelInfo } from "./modelInfo";

export interface ModelsListResponse {
  models: ModelInfo[];
}