# lib/api-zod/src/generated/types/ttsGenerationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Media Suite API
 * OpenAPI spec version: 0.2.0
 */

export interface TtsGenerationRequest {
  modelId: string;
  text: string;
  voice?: string;
  speed?: number;
  /** Reference voice audio URL for voice cloning */
  referenceAudioUrl?: string;
}