# lib/api-zod/src/generated/types/videoGenerationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Media Suite API
 * OpenAPI spec version: 0.2.0
 */

export interface VideoGenerationRequest {
  modelId: string;
  prompt: string;
  negativePrompt?: string;
  duration?: string;
  aspectRatio?: string;
  imageUrl?: string;
  seed?: number;
  generateAudio?: boolean;
}